\name{index}
\alias{index}
\title{chunk indexing}
\description{(internal) index a chunk. This function should not be called by the user directly.}
\usage{index(label, type)}
\arguments{
  \item{label}{the chunk label.}
  \item{type}{the type of chunk to be indexed.}
}

\value{rank of the indexed figr.}

\author{Michael Koohafkan}



\seealso{\code{\link{figr}}, \code{\link{anchors}}}

