\name{kfigr-package}
\alias{kfigr-package}
\alias{kfigr}
\docType{package}
\title{Integrated Code Chunk Anchoring and Referencing for R Markdown Documents}
\description{A streamlined cross-referencing system for R Markdown documents
generated with 'knitr'. R Markdown is  an authoring format for generating
dynamic content from R. It provides a hook for anchoring code
chunks and a function to cross-reference document elements generated from
said chunks, e.g. figures and tables.}
\details{
\tabular{ll}{Package: \tab kfigr\cr
Title: \tab Integrated Code Chunk Anchoring and Referencing for R Markdown Documents\cr
Version: \tab 1.1.0\cr
Date: \tab 2015-03-13\cr
Author: \tab Michael Koohafkan\cr
Maintainer: \tab Michael Koohafkan <michael.koohafkan@gmail.com>\cr
Depends: \tab R (>= 3.1.0)\cr
Imports: \tab knitr (>= 1.6)\cr
Suggests: \tab ggplot2\cr
LazyData: \tab true\cr
VignetteBuilder: \tab knitr\cr
License: \tab GPL (>= 3)\cr
URL: \tab https://github.com/mkoohafkan/kfigr\cr
BugReports: \tab https://github.com/mkoohafkan/kfigr/issues\cr}
}
\author{Michael Koohafkan}

\keyword{ package }


