% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{anchors}
\alias{anchors}
\title{Anchor Information}
\usage{
anchors(tag)
}
\arguments{
\item{tag}{Optional specification of return type. \code{tag = "index"} 
returns a dataframe listing the chunk labels, asigned types and index 
numbers. \code{tag = "history"} returns a dataframe listing all reference 
calls, in order.}
}
\value{
If \code{tag = "index"}, a dataframe listing all anchored chunks. If 
  \code{tag = "history"}, a dataframe listing all references made, in order.
  If the value of \code{tag} matches a specific \code{type}, all references 
  of that \code{type} are provided in a list. If \code{tag} is missing, all 
  references of all types are provided in a nested list.
}
\description{
Retrieves diagnostics such as the anchor index and history. Use for 
code verification and troubleshooting. Also used internally by kfigr.
}
\examples{
figr("foo", type="figure")
figr("bar", type="table")
figr("test", type="figure")
anchors()

}
