% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/figr.R
\name{figr}
\alias{figr}
\title{Code Chunk Reference}
\usage{
figr(label, prefix, link, type)
}
\arguments{
\item{label}{The code chunk label.}

\item{prefix}{Logical: return the prefix (chunk type) along with the rank.}

\item{link}{### Logical: Place an HTML link to the anchor?}

\item{type}{The type of the referenced chunk.}
}
\value{
A text or url reference to the code chunk.
}
\description{
Reference a code chunk.
}
\examples{
figr("test", type="reference")
figr("test")
figr("test", TRUE, link=FALSE)
if(interactive()){
  require(knitr)
  opts_knit$set(kfigr.prefix=TRUE, kfigr.link=TRUE)
  figr("test")
}
}

