% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Chunk Indexing}
\usage{
index(label, type)
}
\arguments{
\item{label}{The chunk label.}

\item{type}{The type of chunk to be indexed.}
}
\value{
The rank of the indexed chunk.
}
\description{
(Internal) index a chunk. Should not be called by the user directly.
}
\seealso{
\code{\link{figr}}, \code{\link{anchors}}
}
