\name{index}
\alias{index}
\title{chunk indexing}
\description{index a chunk with figr}
\usage{index(label, type)}
\arguments{
  \item{label}{the chunk label}
  \item{type}{the type of figr to be indexed}
}
\details{This is an internal function used by kfigr. It is not designed 
to be called by the user directly.}
\value{rank of the indexed figr}

\author{Michael Koohafkan}



\seealso{\code{\link{figr}}, \code{\link{anchors}}}

