% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfigr-package.R
\docType{package}
\name{kfigr-package}
\alias{kfigr-package}
\title{Integrated Code Chunk Anchoring and Referencing for R Markdown Documents}
\description{
A streamlined cross-referencing system for R Markdown documents
generated with 'knitr'. It provides the chunk hook \code{anchor}
and a cross-referencing function \code{figr}. See the vignette to 
get started.
}
