\name{kfigr-package}
\alias{kfigr-package}
\alias{kfigr}
\docType{package}
\title{Integrated chunk anchoring and referencing for knitr documents.}
\description{kfigr is a streamlined version of figr designed for tight integration with knitr. The package defines a hook for anchoring code chunks and the ability to track and cross-reference chunks for producing document elements like figures, tables and code blocks.}
\details{
\tabular{ll}{Package: \tab kfigr\cr
Title: \tab Integrated chunk anchoring and referencing for knitr documents.\cr
Version: \tab 1.0\cr
Author: \tab Michael Koohafkan\cr
Maintainer: \tab Michael Koohafkan <michael.koohafkan@gmail.com>\cr
Depends: \tab R (>= 3.1.0)\cr
Imports: \tab knitr\cr
License: \tab GPL (>=2)\cr
Suggests: \tab bibtex, ggplot2\cr
VignetteBuilder: \tab knitr\cr
LazyData: \tab true\cr}
}
\author{Michael Koohafkan}

\keyword{ package }


