\name{kfigr-package}
\alias{kfigr-package}
\alias{kfigr}
\docType{package}
\title{Integrated chunk anchoring and referencing for knitr documents.}
\description{A streamlined cross-referencing system for R markdown
documents that integrates with knitr. Defines a hook for
anchoring code chunks and a function to track and cross-reference
chunks that produce document elements such as figures and tables.}
\details{
\tabular{ll}{Package: \tab kfigr\cr
Title: \tab Integrated chunk anchoring and referencing for knitr documents.\cr
Version: \tab 1.1.0\cr
Date: \tab 2015-01-16\cr
Author: \tab Michael Koohafkan\cr
Maintainer: \tab Michael Koohafkan <michael.koohafkan@gmail.com>\cr
Depends: \tab R (>= 3.1.0)\cr
Imports: \tab knitr (>= 1.6)\cr
Suggests: \tab ggplot2\cr
LazyData: \tab true\cr
VignetteBuilder: \tab knitr\cr
License: \tab GPL (>= 3)\cr
URL: \tab https://github.com/mkoohafkan/kfigr\cr
BugReports: \tab https://github.com/mkoohafkan/kfigr/issues\cr}
}
\author{Michael Koohafkan}

\keyword{ package }


